Cart not working
up to this point the cart is working to some extent, it adds in one item but the persistence of items in the cart seem like an issue. the cart adds in one item and you cant add more. we also are having issues showing the item list in the cart page. If you go to the home page all memory of the cart is lost.